feat: mint check into cdk
