+++ crates/cdk/src/mint/mod.rs
@@ -29,6 +29,7 @@ mod info;
 mod keysets;
 mod melt;
 mod mint_nut04;
+mod start_up_check;
 mod swap;
 pub mod types;
